+++ Cargo-recent.lock
@@ -114,6 +114,7 @@ dependencies = [
  "bitcoin-io",
  "bitcoin-units",
  "bitcoin_hashes",
+ "hex-conservative",
  "mutagen",
  "ordered",
  "serde",
@@ -180,6 +181,7 @@ source = "registry+https://github.com/rust-lang/crates.io-index"
 checksum = "e1aa273bf451e37ed35ced41c71a5e2a4e29064afb104158f2514bcd71c2c986"
 dependencies = [
  "arrayvec",
+ "serde",
 ]
 
 [[package]]
